+++ biqbin_base.py
@@ -1,163 +1,374 @@
 __version__ = '2.0.0'
 
+import numpy.typing as npt
+
+from typing import Generic, TypeVar
 from abc import ABC, abstractmethod
-import argparse
-from functools import reduce
-from math import gcd
-import numpy as np
-import scipy as sp
-import json
-import warnings
 from glob import glob
+import argparse
+import numpy as np
+import json
 import pyqplib
 
+from utils import from_sparse, check_matrix_validity_wrap, convert_numpy_to_json_serializable
 from biqbin import (run, set_heuristic,
                     default_heuristic,
-                    get_rank, set_read_data,
-                    default_read_data)
+                    get_rank, set_read_data)
 
 
-class DataGetter(ABC):
-    """Base class for all data getters.
+class PrettyPrint:
+    def __str__(self) -> str:
+        return (f'class: {type(self).__name__}')
 
-    Subclasses are responsible for reading a problem instance from some
-    source (file, generator, API, …) and exposing it in a format that can
-    be passed to a solver.
+    def __repr__(self) -> str:
+        return str(self)
 
-    Subclasses must implement:
-        `problem_instance_name` -> str:
-            Returns a human-readable identifier of the problem instance.
-        `problem_instance` -> np.ndarray:
-            Returns the problem instance as a numpy ndarray in the format
-            expected by the solver.
-        `read_file` -> np.ndarray:
-            Reads the raw data (e.g. from disk) and initializes internal
-            state used by ``problem_instance()``.
-    """
-    @abstractmethod
-    def problem_instance_name(self) -> str:
-        """Gets the identifier string of the problem instance.
 
-        Returns:
-            str: name, file path, identifier, ... of the problem instance
-        """
-        ...
+class ProblemMaxCut(PrettyPrint):
+    def __init__(self, maxcut_adjacency_matrix: npt.NDArray[np.floating | np.integer], problem_name: str) -> None:
+        self._maxcut_adjacency_matrix = maxcut_adjacency_matrix
+        self.problem_name = problem_name
 
-    @abstractmethod
-    def problem_instance(self) -> np.ndarray:
-        """Return the problem instance matrix.
+    @property
+    @check_matrix_validity_wrap
+    def maxcut_adjacency_matrix(self):
+        return self._maxcut_adjacency_matrix
 
-        Returns:
-            np.ndarray: A 2D numpy array of shape (n, n)
-        """
-        ...
+    @maxcut_adjacency_matrix.setter
+    def maxcut_adjacency_matrix(self, value: npt.NDArray[np.floating | np.integer]):
+        self._maxcut_adjacency_matrix = value
 
-    @abstractmethod
-    def read_file(self) -> np.ndarray:
-        """Called when Biqbin is ran, reads the problem instance from file
-        and converts it to a numpy ndarray.
+    def __str__(self) -> str:
+        return f'{super().__str__()}\nMaxCut adjacency matrix =\n{self.maxcut_adjacency_matrix}'
 
-        Returns:
-            np.ndarray: A 2D numpy array of shape (n, n)
-        """
-        ...
 
-    def from_sparse(self, sparse_matrix) -> np.ndarray:
-        """Helper function that converts from sparse coo matrix to regular form
+class ProblemQubo(ProblemMaxCut):
+    def __init__(self, Q: np.ndarray, is_minimization: bool, problem_name: str):
+        self.Q: np.ndarray = Q
+        self.is_minimization: bool = is_minimization
+        if self.is_minimization:
+            super().__init__(self.qubo2maxcut(Q), problem_name)
+        else:
+            super().__init__(self.qubo2maxcut(-Q), problem_name)
+
+    @check_matrix_validity_wrap
+    def qubo2maxcut(self, qubo: np.ndarray) -> np.ndarray:
+        """Convert qubo to adjacency matrix that biqbin can read, 
+        optionally optimizes the input data by dividing the values by the gcd.
 
         Args:
-            sparse_matrix (dict): scipy sparse coo matrix
+            qubo (np.ndarray): qubo as 2d numpy array
+        Returns:
+            np.ndarray: adjacency matrix for max cut problem
+        """
+
+        q_sym = 1/2*(qubo.T + qubo)
+
+        q_int = np.array(q_sym, dtype=np.int64)
+        if not np.all(q_sym == q_int):
+            raise ValueError("All QUBO values need to be integers!")
+
+        Qe_plus_c = -np.array([(np.sum(q_sym, 1))])
+        np.fill_diagonal(q_sym, 0)
+
+        return np.block([
+            [q_sym, Qe_plus_c.T],
+            [Qe_plus_c, np.zeros((1, 1))]
+        ])
+
+    def __str__(self) -> str:
+        return f'{super().__str__()}\nQubo =\n{self.Q}'
+
+
+class SolutionMaxCut(PrettyPrint):
+    def __init__(self, biqbin_result: dict) -> None:
+        self.__solution: dict = biqbin_result['maxcut']
+        self.meta_data: dict = biqbin_result['meta_data']
+
+    @property
+    def solution(self):
+        return self.__solution
+
+    def __str__(self) -> str:
+        return (f'{super().__str__()}\n'
+                f'--- Max-Cut ---\n'
+                f' Computed value = {self.__solution['computed_val']}\n'
+                f'Solution MaxCut = {self.__solution['solution']}\n'
+                f'              x = {self.__solution['x']}\n')
+
+
+class SolutionQubo(SolutionMaxCut):
+    def __init__(self, biqbin_result: dict, problem: ProblemQubo) -> None:
+        super().__init__(biqbin_result)
+        self.problem = problem
+        self.solution_maxcut = super().solution
+        self.__solution = self.maxcut2qubo()
+
+    @property
+    def solution(self):
+        return self.__solution
+
+    def maxcut2qubo(self) -> dict:
+        """Write qubo solution from the retrievied maxcut solution
 
         Returns:
-            np.ndarray: regular form matrix
+            dict: Qubo specific result (computed_val, x, solution, cardinality, obj)
         """
-        return sp.sparse.coo_matrix(
-            (sparse_matrix['data'],
-             (sparse_matrix['row'], sparse_matrix['col'])),
-            shape=sparse_matrix['shape'], dtype='float'
-        ).todense().getA()
+        qubo_solution, qubo_x = self._maxcut_solution2qubo_solution(
+            self.solution_maxcut["solution"]
+        )
+
+        computed_val = float(self.problem.Q.dot(qubo_x).dot(qubo_x))
+        return {'computed_val': computed_val,
+                'solution': qubo_solution,
+                'x': qubo_x,
+                'cardinality': float(sum(qubo_x)),
+                'minimization': self.problem.is_minimization
+                }
+
+    def _maxcut_solution2qubo_solution(self, maxcut_solution: np.ndarray):
+        """Convert maxcut solution nodes to qubo solution nodes
+
+        Args:
+            maxcut_solution (np.ndarray): maxcut solution found by biqbin
+        Returns:
+            np.ndarray: qubo solution nodes
+        """
+
+        n, _ = self.problem.Q.shape
+
+        _x_mc = np.array(maxcut_solution, dtype=int)-1
+        x_mc_sol = np.ones(n + 1)
+        xx = np.zeros(n + 1, dtype=int)
+        xx[_x_mc] = 1
+
+        x_mc_sol[_x_mc] = -1
+        x_mc_sol *= -x_mc_sol[-1]
+        y = 1/2*(x_mc_sol+1)[:-1]
+        qubo_solution = np.nonzero(y)[0] + 1
+
+        return qubo_solution.tolist(), y.astype(int).tolist()
+
+    def __str__(self):
+        return (
+            f'{super().__str__()}\n'
+            f'--- QUBO ---\n'
+            f' Computed value = {self.__solution['computed_val']}\n'
+            f'       Solution = {self.__solution['solution']}\n'
+            f'              x = {self.__solution['x']}\n'
+            f'is_minimization = {self.problem.is_minimization}\n'
+        )
 
 
-class DataGetterMaxCutDefault(DataGetter):
-    """
-    Uses the default C implementation or MaxCut, reads and parses maxcut instance file in edge weight list format and parses
-    into the adjacency matrix.
-    """
-
-    def __init__(self, filename: str):
+class FromFile(ABC, PrettyPrint):
+    def __init__(self, filename: str, problem_name: str | None = None):
         self.filename: str = filename
-        self.adj_matrix: np.ndarray | None = None
+        self.problem_name: str = problem_name if problem_name else filename
 
-    def problem_instance_name(self) -> str:
-        """Get the instance file path
+    @abstractmethod
+    def read(self) -> ProblemMaxCut:
+        ...
 
-        Returns:
-            str: path to instance file
-        """
-        return self.filename
-
-    def problem_instance(self) -> np.ndarray:
-        """Gets the adjacency matrix from the instance file
-        """
-        if isinstance(self.adj_matrix, np.ndarray):
-            return self.adj_matrix
-        else:
-            raise ValueError(
-                f"Expected self.adj_matrix to be of type 'numpy.ndarray', but got '{type(self.adj_matrix)}'.")
-
-    def read_file(self) -> np.ndarray:
-        self.adj_matrix = default_read_data(self.filename)
-        if isinstance(self.adj_matrix, np.ndarray):
-            return self.adj_matrix
-        else:
-            raise ValueError(
-                f"Expected self.adj_matrix to be of type 'numpy.ndarray', but got '{type(self.adj_matrix)}'.")
+    def __str__(self):
+        return (
+            f'{super().__str__()}\n'
+            f'Input filename: {self.filename}')
 
 
-class DataGetterAdjacencyJson(DataGetterMaxCutDefault):
-    """
-    DataGetter for the Maxcut class, reads and parses json serialized dict with adj key and sparse coo matrix as value.
-    """
+TSolution = TypeVar('TSolution', bound=SolutionMaxCut)
 
-    def read_file(self) -> np.ndarray:
+
+class ToFile(ABC, PrettyPrint, Generic[TSolution]):
+    def __init__(self, filename: str, overwrite: bool = False):
+        self.filename: str = filename
+        self.overwrite: bool = overwrite
+
+    @abstractmethod
+    def write(self, solution: TSolution) -> None:
+        ...
+
+    def get_output_path(self, out_file: str, overwrite: bool) -> str:
+        out_file = out_file[:-5] if out_file.endswith('.json') else out_file
+        if overwrite:
+            return out_file
+
+        file_count = len(glob(f'{out_file}*.json'))
+        if file_count > 0:
+            out_file += f'_{file_count}'
+        return out_file + '.json'
+
+    def __str__(self):
+        return (
+            f'{super().__str__()}\n'
+            f'Output filename: {self.filename}\n'
+            f'Overwrite output: {self.overwrite}'
+        )
+
+
+class MaxCutFromJson(FromFile):
+    def read(self) -> ProblemMaxCut:
         with open(self.filename, "r") as f:
             mc_data = json.load(f)
 
-        self.adj_matrix = self.from_sparse(mc_data["adjacency"])
+        adj_matrix = from_sparse(mc_data["adjacency"])
+        return ProblemMaxCut(adj_matrix, self.filename)
 
-        adj_int = np.array(self.adj_matrix, dtype=np.int64)
-        if not np.all(self.adj_matrix == adj_int):
+
+class MaxCutFromEdgeWeights(FromFile):
+    def read(self) -> ProblemMaxCut:
+        with open(self.filename, 'r') as f:
+            # Read number of vertices and edges
+            num_vertices, num_edges = map(int, f.readline().split())
+            adj_matrix = np.zeros(
+                (num_vertices, num_vertices), dtype=np.float64)
+
+            for _ in range(num_edges):
+                i, j, weight = f.readline().split()
+                i, j = int(i) - 1, int(j) - 1  # Convert to zero-based indexing
+                weight = float(weight)
+
+                adj_matrix[i, j] = weight
+                adj_matrix[j, i] = weight
+            return ProblemMaxCut(adj_matrix, self.problem_name)
+
+
+class MaxCutToJson(ToFile):
+    def write(self, solution: SolutionMaxCut, with_metadata=True) -> None:
+        """Save the solution as JSON file.
+
+        Args:
+            solution (SolutionMaxCut): Solution class returned by Biqbin after solving the problem
+            with_metadata (bool, optional): Add meta_data to output file. Defaults to True.
+        """
+
+        # Check if output filename exists if we are not overriding and replace with filename_N.json
+        output_path = self.get_output_path(self.filename, self.overwrite)
+
+        save_output = {
+            'maxcut': solution.solution
+        }
+        if with_metadata:
+            save_output['meta_data'] = solution.meta_data
+
+        with open(output_path, 'w') as f:
+            json.dump(save_output, f,
+                      default=convert_numpy_to_json_serializable)
+
+
+class QuboToJson(ToFile):
+    """Helper class to save QUBO solution as json file.
+
+    Args:
+        ToFile (_type_): _description_
+    """
+
+    def write(self, solution: SolutionQubo, with_metadata=True) -> None:
+        """Save qubo solution to a json file.
+
+        Args:
+            solution (SolutionQubo): Solution class returned by Biqbin after solving the problem
+            with_metadata (bool, optional): Add meta_data to output. Defaults to True.
+            with_maxcut_solution (bool, optional): Add MaxCut solution to save output. Defaults to False.
+        """
+        # Check if output filename exists if we are not overriding and replace with filename_N.json
+        output_path = self.get_output_path(self.filename, self.overwrite)
+
+        save_output = {
+            'qubo': solution.solution
+        }
+        save_output['maxcut'] = solution.solution_maxcut
+        if with_metadata:
+            save_output['meta_data'] = solution.meta_data
+
+        with open(output_path, 'w') as f:
+            json.dump(save_output, f,
+                      default=convert_numpy_to_json_serializable)
+
+
+class QuboFromJson(FromFile):
+    """Reads qubo instance file, should be a json dictionary with "qubo" key
+    and a COO sparse matrix with data, row and col.
+    """
+
+    def read(self) -> ProblemQubo:
+        with open(self.filename, "r") as f:
+            qubo_data = json.load(f)
+
+        qubo = from_sparse(qubo_data["qubo"])
+        return ProblemQubo(qubo, True, self.filename)
+
+
+class QuboFromQPLIB(FromFile):
+    """DataGetter for QPLIB instances https://qplib.zib.de/, 
+    only unconstrained binary problems are allowed.
+    """
+
+    def read(self) -> ProblemQubo:
+        """Reads .qplib format and constructs a qubo. 
+        Checks if the problem itself is valid for Biqbin solver, while the integer check 
+        is done when converting the constructed QUBO to Max-Cut form.
+
+        Raises:
+            ValueError: Only unconstrained problems are valid
+            ValueError: Only binary problems are valid
+
+        Returns:
+            ProblemQUBO
+        """
+        qplib_problem = pyqplib.read_problem(self.filename)
+
+        # Check if reading qplib format worked
+        if not isinstance(qplib_problem, pyqplib.Problem):
             raise ValueError(
-                "All values in the adjacency matrix need to be integers!")
+                f"Failed reading qplib problem at path {self.filename}!")
 
-        return self.adj_matrix
+        # Check if the problem fits the solver
+        if qplib_problem.description.cons_type != pyqplib.ProblemConsType.UNCONSTRAINED:
+            raise ValueError("Biqbin can only handle unconstrained problems!")
+        if qplib_problem.description.var_type != pyqplib.ProblemVarType.BINARY:
+            raise ValueError("Problem is not binary!")
+
+        # pyqplib has it's own matrix represantation
+        qubo = qplib_problem.obj.mat.full().todense().T
+
+        qubo = np.triu(qubo) / 2
+        qubo += np.diag(qplib_problem.obj.lin)
+
+        # Update goal of the objective function
+        minimize = qplib_problem.obj.sense == pyqplib.Sense.MINIMIZE
+        return ProblemQubo(qubo, minimize, self.problem_name)
 
 
-class MaxCutSolver:
+TProblem = TypeVar("TProblem", bound="ProblemMaxCut")
+
+
+class MaxCutSolver(Generic[TProblem], PrettyPrint):
     """Default MaxCut Biqbin Wrapper, runs Biqbin MaxCut using its original functions
     """
     solver_name = f'PyBiqBin-MaxCut {__version__}'
 
-    def __init__(self, data_getter: DataGetter, params: str, time_limit: int = 0):
+    def __init__(self, problem: TProblem, params: str, time_limit: int = 0):
         """Initialize the solver
 
         Args:
-            problem_instance_name (str): path to problem instance in edge weight list format
+            problem (ProblemMaxCut): MaxCut problem
             params (str): path to parameters file
+            time_limit (int): time limit in seconds
         """
-        self.data_getter: DataGetter = data_getter
+        self.problem: TProblem = problem
         self.params = params
+        self.time_limit = time_limit
         set_read_data(self.read_data)
         set_heuristic(self.heuristic)
-        self.time_limit = time_limit
 
+    @check_matrix_validity_wrap
     def read_data(self) -> np.ndarray:
-        """Transform edge weight list into an adjacancy matrix
+        """Return the maxcut_adjacency_matrix from Problem
 
         Returns:
             np.ndarray: adjacency matrix
         """
-        return self.data_getter.read_file()
+        return self.problem.maxcut_adjacency_matrix
 
     def heuristic(self, L0: np.ndarray, L: np.ndarray, xfixed: np.ndarray, sol_X: np.ndarray, x: np.ndarray) -> float:
         """Default heuristic (heuristic_unpacked in heuristic.c)
@@ -174,19 +385,24 @@ class MaxCutSolver:
         """
         return default_heuristic(L0, L, xfixed, sol_X, x)
 
-    def run(self):
+    def run(self) -> SolutionMaxCut | None:
         """Runs Biqbin Maxcut solver
 
         Returns:
             dict: result dict with keys: "max_val" - max cut solution value, "solution" - nodes in this solution, "time" - spent solving 
         """
-        result = run(self.solver_name, self.data_getter.problem_instance_name(
-        ), self.params, self.time_limit)
+        biqbin_result = run(self.solver_name, self.problem.problem_name,
+                            self.params, self.time_limit)
         if (self.get_rank() == 0):
-            result['meta_data']['instance'] = self.data_getter.problem_instance_name()
-            result['meta_data']['parameters'] = {
-                'time_limit': self.time_limit if self.time_limit > 0 else None}
-            return result
+            if biqbin_result is None:
+                raise ValueError(
+                    'result from biqbin is None, computation failed!')
+
+            biqbin_result['meta_data']['instance'] = self.problem.problem_name
+            biqbin_result['meta_data']['parameters'] = {
+                'time_limit': self.time_limit if self.time_limit > 0 else None
+            }
+            return SolutionMaxCut(biqbin_result)
         else:
             return None
 
@@ -198,291 +414,33 @@ class MaxCutSolver:
         """
         return get_rank()
 
-    def save_result(self, result, output_path_in=None, overwrite=False):
-        """Save the result dictionary as JSON file.
 
-        Args:
-            result (dict): result dictionary returned by biqbin
-            output_path (str, optional): custom path to an output file. Defaults to None.
-        """
-        if not output_path_in:
-            output_path = self.data_getter.problem_instance_name() + '.output'
-        else:
-            output_path = output_path_in
-
-        # Always overwrite if filepath specific
-        if not overwrite and not output_path_in:
-            file_count = len(glob(f'{output_path}*.json'))
-            if file_count > 0:
-                output_path += f'_{file_count}'
-        if not output_path_in:
-            output_path += '.json'
-        with open(output_path, "w") as f:
-            json.dump(result, f, default=self._convert_numpy)
-
-    def _convert_numpy(self, obj):
-        if isinstance(obj, np.ndarray):
-            return obj.tolist()
-        if isinstance(obj, np.generic):
-            return obj.item()
-        raise TypeError(f"Object of type {type(obj)} is not JSON serializable")
-
-
-class DataGetterQUBO(DataGetter):
-    """Abstract class for QUBO DataGetters from which all DataGetters for QUBOSolver must inherit.
-    It is responsible for providing the solver with the qubo in upper triangular np.ndarray and 
-
-    Subclasses must implement:
-        `problem_instance_name` -> str:
-            Returns a human-readable identifier of the problem instance.
-        `problem_instance` -> np.ndarray:
-            Returns the problem instance as a numpy ndarray in the format
-            expected by the solver.
-        `read_file` -> np.ndarray:
-            Reads the raw data (e.g. from disk) and initializes internal
-            state used by ``problem_instance()``.
-        `minimize` -> bool:
-            return True if the objective is to minimize, False if maximize.
-    """
-
-    @abstractmethod
-    def minimize(self) -> bool:
-        """Defines if the problem is to minimize or maximize the objective function.
-
-        Returns:
-            bool: True - minimize, False - maximize
-        """
-        ...
-
-    def get_qubo4biqbin(self) -> np.ndarray:
-        """Wrapper for read_file() method, calls `read_file` and 
-        returns -qubo if the objective is to maximize.
-
-        Returns:
-            np.ndarray: qubo if minimize(), else -qubo
-        """
-        qubo = self.read_file()
-        return qubo if self.minimize() else -qubo
-
-    def get_qubo_result(self, result: dict) -> dict:
-        """Add 'qubo' result, built from the 'maxcut' result dict biqbin returned
-
-        Args:
-            result (dict): Biqbin result dictionary containing maxcut results
-
-        Returns:
-            dict: Qubo specific result (computed_val, x, solution, cardinality, obj)
-        """
-        qubo_solution, qubo_x = self._maxcut_solution2qubo_solution(
-            result["maxcut"]["solution"]
-        )
-
-        computed_val = float(self.problem_instance().dot(qubo_x).dot(qubo_x))
-        return {'computed_val': computed_val,
-                'solution': qubo_solution,
-                'x': qubo_x,
-                'cardinality': float(sum(qubo_x)),
-                'obj': 'minimize' if self.minimize() else 'maximize'
-                }
-
-    def _maxcut_solution2qubo_solution(self, maxcut_solution: np.ndarray):
-        """Convert maxcut solution nodes to qubo solution nodes
-
-        Args:
-            maxcut_solution (np.ndarray): maxcut solution found by biqbin
-        Returns:
-            np.ndarray: qubo solution nodes
-        """
-
-        n, _ = self.problem_instance().shape
-
-        _x_mc = np.array(maxcut_solution, dtype=int)-1
-        x_mc_sol = np.ones(n + 1)
-        xx = np.zeros(n + 1, dtype=int)
-        xx[_x_mc] = 1
-
-        x_mc_sol[_x_mc] = -1
-        x_mc_sol *= -x_mc_sol[-1]
-        y = 1/2*(x_mc_sol+1)[:-1]
-        qubo_solution = np.nonzero(y)[0] + 1
-
-        return qubo_solution.tolist(), y.astype(int).tolist()
-
-
-class DataGetterJson(DataGetterQUBO):
-    """Reads qubo instance file, should be a json dictionary with "qubo" key
-    and a COO sparse matrix with data, row and col. Indices starts from zero.
-    """
-
-    def __init__(self, filename: str):
-        """Load data from json file and save the data and qubo
-
-        Args:
-            filename (str): path to file
-        """
-        self.filename: str = filename
-        self.qubo: np.ndarray | None = None
-
-    def problem_instance_name(self) -> str:
-        """Get the instance file path
-
-        Returns:
-            str: path to instance file
-        """
-        return self.filename
-
-    def problem_instance(self) -> np.ndarray:
-        """Gets the qubo in upper triangular form
-
-        Returns:
-            nd.ndarray: qubo
-        """
-        if isinstance(self.qubo, np.ndarray):
-            return self.qubo
-        else:
-            raise ValueError(
-                f'Expected self.qubo to be of type numpy.ndarray, got {type(self.qubo)}.')
-
-    def read_file(self) -> np.ndarray:
-        with open(self.filename, "r") as f:
-            self.qubo_data = json.load(f)
-
-        self.qubo = self.from_sparse(self.qubo_data["qubo"])
-        return self.qubo
-
-    def minimize(self) -> bool:
-        """DataGetterJson reads minimization problems by default
-
-        Returns:
-            bool: True - minimize
-        """
-        return True
-
-
-class DataGetterQPLIB(DataGetterQUBO):
-    """DataGetter for QPLIB instances https://qplib.zib.de/, 
-    only unconstrained binary problems are allowed. 
-
-    Args:
-        DataGetterQUBO (class): inherits from DataGetterQUBO, the default qubo datagetter
-    """
-
-    def __init__(self, filename: str):
-        self.filename: str = filename
-        self._minimize: bool = True
-        self.qubo: np.ndarray | None
-        self.pqlib_problem: pyqplib.desc.ProblemDescription | pyqplib.Problem | None = None
-
-    def read_file(self):
-        """Reads .qplib format and constructs a qubo. 
-        Checks if the problem itself is valid for Biqbin solver, while the integer check 
-        is done when converting the constructed QUBO to Max-Cut form.
-
-        Raises:
-            ValueError: Only unconstrained problems are valid
-            ValueError: Only binary problems are valid
-
-        Returns:
-            np.ndarray: constructed qubo in upper triangular form
-        """
-        self.pqlib_problem = pyqplib.read_problem(self.filename)
-
-        # Check if the problem fits the solver
-        if self.pqlib_problem.description.cons_type != pyqplib.ProblemConsType.UNCONSTRAINED:
-            raise ValueError("Biqbin can only handle unconstrained problems!")
-        if self.pqlib_problem.description.var_type != pyqplib.ProblemVarType.BINARY:
-            raise ValueError("Problem is not binary!")
-
-        # pyqplib has it's own matrix represantation
-        self.qubo = self.pqlib_problem.obj.mat.full().todense().T
-
-        self.qubo = np.triu(self.qubo) / 2
-        self.qubo += np.diag(self.pqlib_problem.obj.lin)
-
-        # Update goal of the objective function
-        self._minimize = self.pqlib_problem.obj.sense == pyqplib.Sense.MINIMIZE
-        return self.qubo
-
-    def minimize(self) -> bool:
-        return self._minimize
-
-    def problem_instance(self) -> np.ndarray:
-        if isinstance(self.qubo, np.ndarray):
-            return self.qubo
-        else:
-            raise ValueError(f'Expected self.qubo to be of type numpy.ndarray, got {type(self.qubo)}.')
-
-    def problem_instance_name(self) -> str:
-        return self.filename
-
-
-class QUBOSolver(MaxCutSolver):
+class QUBOSolver(MaxCutSolver[ProblemQubo]):
     solver_name = f'PyBiqBin-QUBO {__version__}'
 
-    def __init__(self, data_getter: DataGetterQUBO, params: str, optimize_input: bool = False, time_limit: int = 0):
-        super().__init__(data_getter, params, time_limit)
-        self.optimize_input: bool = optimize_input
-        self.gcd: int = 1
+    def __init__(self, problem: ProblemQubo, params: str, time_limit: int = 0):
+        super().__init__(problem, params, time_limit)
 
-    def read_data(self) -> np.ndarray:
-        """Read qubo json file, return an adjacency matrix for maxcut
-
-        Returns:
-            np.ndarray: adjacency matrix
-        """
-
-        return self._qubo2maxcut(self.data_getter.get_qubo4biqbin())
-
-    def run(self) -> dict | None:
+    def run(self) -> SolutionQubo | None:
         """Runs the original biqbin then adds the qubo solution nodes to the result dict
 
         Returns:
             dict: result dict containing "maxcut" and "qubo" keys with their respective solutions
         """
-        result = super().run()
+        biqbin_result = run(self.solver_name, self.problem.problem_name,
+                            self.params, self.time_limit)
         if self.get_rank() == 0:
-            if result is None:
+            if biqbin_result is None:
                 raise ValueError("result is None, solution not retrieved!")
 
-            result['maxcut']['computed_val'] *= self.gcd
-            result['meta_data']['parameters']['optimize_input'] = self.optimize_input
-            result['meta_data']['parameters']['gcd'] = self.gcd
+            biqbin_result['meta_data']['instance'] = self.problem.problem_name
+            biqbin_result['meta_data']['parameters'] = {
+                'time_limit': self.time_limit if self.time_limit > 0 else None
+            }
 
-            result['qubo'] = self.data_getter.get_qubo_result(result)
-            return result
+            return SolutionQubo(biqbin_result, self.problem)
         else:
             return None
-
-    def _qubo2maxcut(self, qubo: np.ndarray) -> np.ndarray:
-        """Convert qubo to adjacency matrix that biqbin can read, 
-        optionally optimizes the input data by dividing the values by the gcd.
-
-        Args:
-            qubo (np.ndarray): qubo as 2d numpy array
-        Returns:
-            np.ndarray: adjacency matrix for max cut problem
-        """
-
-        q_sym = 1/2*(qubo.T + qubo)
-
-        q_int = np.array(q_sym, dtype=np.int64)
-        if not np.all(q_sym == q_int):
-            raise ValueError("All QUBO values need to be integers!")
-
-        if self.optimize_input:
-            unique_fields = set(q_sym.astype(int).flatten())
-            greatest_common_divisor = reduce(gcd, unique_fields)
-            if greatest_common_divisor > 1:
-                q_sym /= greatest_common_divisor
-                self.gcd = greatest_common_divisor
-
-        Qe_plus_c = -np.array([(np.sum(q_sym, 1))])
-        np.fill_diagonal(q_sym, 0)
-
-        return np.block([
-            [q_sym, Qe_plus_c.T],
-            [Qe_plus_c, np.zeros((1, 1))]
-        ])
 
 
 class BaseParser(argparse.ArgumentParser):
