+++ biqbin_maxcut.py
@@ -1,5 +1,5 @@
 import argparse
-from biqbin_base import MaxCutSolver, DataGetterMaxCutDefault, DataGetterAdjacencyJson, ParserMaxCut
+from biqbin_base import MaxCutSolver, MaxCutFromJson, MaxCutFromEdgeWeights, ParserMaxCut, MaxCutToJson
 
 """
     Default MaxCut Biqbin wrapper example
@@ -12,14 +12,18 @@ if __name__ == '__main__':
     
     # Create an instance of the MaxCutSolver passing in the above arguments
     if args.edge_weight:
-        data_getter = DataGetterMaxCutDefault(args.problem_instance)
+        file_loader = MaxCutFromEdgeWeights(args.problem_instance)
     else:
-        data_getter = DataGetterAdjacencyJson(args.problem_instance)
-    solver = MaxCutSolver(data_getter, args.params, args.time)
-    result = solver.run()  # run the solver
+        file_loader = MaxCutFromJson(args.problem_instance)
+    
+    problem = file_loader.read()
+    solver = MaxCutSolver(problem, args.params, args.time)
+    solution = solver.run()
 
     rank = solver.get_rank()
     if rank == 0:
         # Print the results if master rank
-        print(result)
-        solver.save_result(result, output_path_in=args.output, overwrite=args.overwrite)
+        if solution is None:
+            raise ValueError(f'Solution to problem {problem} not found!')
+        print(solution)
+        MaxCutToJson(args.problem_instance + '.output.json', args.overwrite).write(solution)
