+++ tests/qplib/5881.qplib-expected_output.json
@@ -1,1 +1,1 @@
-{"time": 7.416342834, "meta_data": {"time": 7.416342834, "time_limit_reached": false, "eval_bab_nodes": 3, "heuristic_run_count": 61, "num_workers_used": 2, "instance": "tests/qplib/5881.qplib", "parameters": {"time_limit": null, "optimize_input": false, "gcd": 1}}, "maxcut": {"computed_val": 13067.0, "solution": [1, 3, 4, 5, 8, 9, 10, 11, 12, 13, 14, 15, 17, 20, 21, 22, 23, 25, 28, 30, 31, 33, 34, 35, 37, 38, 40, 41, 44, 45, 48, 49, 51, 52, 54, 56, 58, 59, 60, 61, 63, 64, 66, 68, 70, 71, 72, 74, 75, 76, 77, 78, 81, 82, 84, 86, 87, 90, 91, 93, 95, 96, 97, 98, 99, 100, 102, 105, 106, 107, 108, 111, 114, 116, 117, 118, 119, 120], "x": [1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0]}, "qubo": {"computed_val": 13067.0, "solution": [1, 3, 4, 5, 8, 9, 10, 11, 12, 13, 14, 15, 17, 20, 21, 22, 23, 25, 28, 30, 31, 33, 34, 35, 37, 38, 40, 41, 44, 45, 48, 49, 51, 52, 54, 56, 58, 59, 60, 61, 63, 64, 66, 68, 70, 71, 72, 74, 75, 76, 77, 78, 81, 82, 84, 86, 87, 90, 91, 93, 95, 96, 97, 98, 99, 100, 102, 105, 106, 107, 108, 111, 114, 116, 117, 118, 119, 120], "x": [1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1], "cardinality": 78.0, "obj": "maximize"}}
+{"meta_data": {"time": 7.416342834, "time_limit_reached": false, "eval_bab_nodes": 3, "heuristic_run_count": 61, "num_workers_used": 2, "instance": "tests/qplib/5881.qplib", "parameters": {"time_limit": null, "optimize_input": false, "gcd": 1}}, "maxcut": {"computed_val": 13067.0, "solution": [1, 3, 4, 5, 8, 9, 10, 11, 12, 13, 14, 15, 17, 20, 21, 22, 23, 25, 28, 30, 31, 33, 34, 35, 37, 38, 40, 41, 44, 45, 48, 49, 51, 52, 54, 56, 58, 59, 60, 61, 63, 64, 66, 68, 70, 71, 72, 74, 75, 76, 77, 78, 81, 82, 84, 86, 87, 90, 91, 93, 95, 96, 97, 98, 99, 100, 102, 105, 106, 107, 108, 111, 114, 116, 117, 118, 119, 120], "x": [1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1, 0]}, "qubo": {"computed_val": 13067.0, "solution": [1, 3, 4, 5, 8, 9, 10, 11, 12, 13, 14, 15, 17, 20, 21, 22, 23, 25, 28, 30, 31, 33, 34, 35, 37, 38, 40, 41, 44, 45, 48, 49, 51, 52, 54, 56, 58, 59, 60, 61, 63, 64, 66, 68, 70, 71, 72, 74, 75, 76, 77, 78, 81, 82, 84, 86, 87, 90, 91, 93, 95, 96, 97, 98, 99, 100, 102, 105, 106, 107, 108, 111, 114, 116, 117, 118, 119, 120], "x": [1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 1], "cardinality": 78.0, "obj": "maximize"}}
