+++ tests/qplib/kcluster40_025_10_1.qplib-expected_output.json
@@ -1,1 +1,1 @@
-{"time": 0.04112426899999999, "meta_data": {"time": 0.04112426899999999, "time_limit_reached": false, "eval_bab_nodes": 1, "heuristic_run_count": 5, "num_workers_used": 0, "instance": "tests/qplib/kcluster40_025_10_1.qplib", "parameters": {"time_limit": null, "optimize_input": false, "gcd": 1}}, "maxcut": {"computed_val": 3900.0, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0]}, "qubo": {"computed_val": -3900.0, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1], "cardinality": 10.0, "obj": "minimize"}}
+{"meta_data": {"time": 0.04112426899999999, "time_limit_reached": false, "eval_bab_nodes": 1, "heuristic_run_count": 5, "num_workers_used": 0, "instance": "tests/qplib/kcluster40_025_10_1.qplib", "parameters": {"time_limit": null, "optimize_input": false, "gcd": 1}}, "maxcut": {"computed_val": 3900.0, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0]}, "qubo": {"computed_val": -3900.0, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1], "cardinality": 10.0, "obj": "minimize"}}
