+++ tests/qplib/kcluster80_025_20_1.qplib-expected_output.json
@@ -1,1 +1,1 @@
-{"time": 16.605838744000003, "meta_data": {"time": 16.605838744000003, "time_limit_reached": false, "eval_bab_nodes": 35, "heuristic_run_count": 502, "num_workers_used": 2, "instance": "tests/qplib/kcluster80_025_20_1.qplib", "parameters": {"time_limit": null, "optimize_input": false, "gcd": 1}}, "maxcut": {"computed_val": 16400.0, "solution": [2, 5, 9, 13, 14, 17, 23, 24, 32, 33, 38, 43, 46, 47, 50, 54, 56, 63, 65, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0]}, "qubo": {"computed_val": -16400.0, "solution": [2, 5, 9, 13, 14, 17, 23, 24, 32, 33, 38, 43, 46, 47, 50, 54, 56, 63, 65, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1], "cardinality": 20.0, "obj": "minimize"}}
+{"meta_data": {"time": 16.605838744000003, "time_limit_reached": false, "eval_bab_nodes": 35, "heuristic_run_count": 502, "num_workers_used": 2, "instance": "tests/qplib/kcluster80_025_20_1.qplib", "parameters": {"time_limit": null, "optimize_input": false, "gcd": 1}}, "maxcut": {"computed_val": 16400.0, "solution": [2, 5, 9, 13, 14, 17, 23, 24, 32, 33, 38, 43, 46, 47, 50, 54, 56, 63, 65, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0]}, "qubo": {"computed_val": -16400.0, "solution": [2, 5, 9, 13, 14, 17, 23, 24, 32, 33, 38, 43, 46, 47, 50, 54, 56, 63, 65, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1], "cardinality": 20.0, "obj": "minimize"}}
