+++ tests/qubos/40/kcluster40_025_10_1.json-expected_output.json
@@ -1,1 +1,1 @@
-{"time": 0.04199293900000001, "maxcut": {"computed_val": 3899.999999999999, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0]}, "qubo": {"computed_val": -3900.0, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1], "cardinality": 10.0}}
+{"meta_data": {"time": 0.03988125000000009, "time_limit_reached": false, "eval_bab_nodes": 1, "heuristic_run_count": 5, "num_workers_used": 0, "instance": "tests/qubos/40/kcluster40_025_10_1.json", "parameters": {"time_limit": null, "optimize_input": false, "gcd": 1}}, "maxcut": {"computed_val": 3900.0, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0]}, "qubo": {"computed_val": -3900.0, "solution": [1, 3, 4, 9, 27, 32, 34, 38, 39, 40], "x": [1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1], "cardinality": 10.0, "obj": "minimize"}}
