+++ tests/qubos/80/kcluster80_025_20_1.json-expected_output.json
@@ -1,1 +1,1 @@
-{"time": 9.678719459, "maxcut": {"computed_val": 16400.0, "solution": [2, 5, 9, 13, 15, 17, 24, 26, 31, 33, 38, 47, 50, 53, 54, 56, 63, 70, 78, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0]}, "qubo": {"computed_val": -16400.0, "solution": [2, 5, 9, 13, 15, 17, 24, 26, 31, 33, 38, 47, 50, 53, 54, 56, 63, 70, 78, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1], "cardinality": 20.0}}
+{"meta_data": {"time": 9.678719459}, "maxcut": {"computed_val": 16400.0, "solution": [2, 5, 9, 13, 15, 17, 24, 26, 31, 33, 38, 47, 50, 53, 54, 56, 63, 70, 78, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0]}, "qubo": {"computed_val": -16400.0, "solution": [2, 5, 9, 13, 15, 17, 24, 26, 31, 33, 38, 47, 50, 53, 54, 56, 63, 70, 78, 80], "x": [0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1], "cardinality": 20.0}}
